+++ Gems/EMotionFX/Code/MysticQt/Source/KeyboardShortcutManager.cpp
@@ -163,7 +163,7 @@ namespace MysticQt
         // iterate through the groups and save all actions for them
         for (const AZStd::unique_ptr<Group>& group : m_groups)
         {
-            settings->beginGroup(QString::fromUtf8(group->GetName().data(), group->GetName().size()));
+            settings->beginGroup(QString::fromUtf8(group->GetName().data(), static_cast<int>(group->GetName().size())));
 
             // iterate through the actions and save them
             for (const AZStd::unique_ptr<Action>& action : group->GetActions())
