+++ Gems/PhysXDebug/Code/Source/SystemComponent.cpp
@@ -561,7 +561,7 @@ namespace PhysXDebug
 
     static void physx_CullingBoxSize([[maybe_unused]] const AZ::ConsoleCommandContainer& arguments)
     {
-        const int argumentCount = arguments.size();
+        const size_t argumentCount = arguments.size();
         if (argumentCount == 1)
         {
             float newCullingBoxSize = (float)strtol(AZ::CVarFixedString(arguments[0]).c_str(), nullptr, 10);
@@ -577,7 +577,7 @@ namespace PhysXDebug
 
     static void physx_Debug([[maybe_unused]] const AZ::ConsoleCommandContainer& arguments)
     {
-        const int argumentCount = arguments.size();
+        const size_t argumentCount = arguments.size();
 
         if (argumentCount == 1)
         {
